4.1-5.1 lessons on cplusplus.com

+++ LearnCPPConsoleApplication1/LearnCPPConsoleApplication1/PowerCalculator.h
@@ -4,8 +4,8 @@ class PowerCalculator
 public:
 	PowerCalculator();
 	~PowerCalculator();
-	static int powByRecursion(int base, int exp);
-	static int powByLoop(int base, int exp);
+	static double powByRecursion(int base, int exp);
+	static double powByLoop(int base, int exp);
 	static int powByExponentialBySquaring(int base, int exp);
 };
